+++ src/indicators/atr.rs
@@ -4,10 +4,10 @@ use crate::indicators::{Value,Price, Indicator};
 #[derive(Clone, Debug)]
 pub struct Atr {
     periods: u32,
-    value: Option<f32>,
-    prev_value: Option<f32>,
-    prev_close: Option<f32>,
-    warmup_trs: Vec<f32>,
+    value: Option<f64>,
+    prev_value: Option<f64>,
+    prev_close: Option<f64>,
+    warmup_trs: Vec<f64>,
 }
 
 
@@ -25,8 +25,8 @@ impl Atr{
         }
     }
 
-    pub fn normalized(&self, price: f32) -> Option<Value> {
-    if price.abs() < f32::EPSILON {
+    pub fn normalized(&self, price: f64) -> Option<Value> {
+    if price.abs() < f64::EPSILON {
         return None;
     }
     self.value.map(|value| Value::AtrValue((value / price) * 100.0))
@@ -55,13 +55,13 @@ impl Indicator for Atr{
     if self.value.is_none() {
         self.warmup_trs.push(tr);
         if self.warmup_trs.len() == self.periods as usize {
-            let sum: f32 = self.warmup_trs.iter().sum();
-            let initial_atr = sum / self.periods as f32;
+            let sum: f64 = self.warmup_trs.iter().sum();
+            let initial_atr = sum / self.periods as f64;
             self.value = Some(initial_atr);
             self.prev_value = Some(initial_atr);
         }
     } else if let Some(prev_atr) = self.value {
-        let new_atr = (prev_atr * (self.periods as f32 - 1.0) + tr) / self.periods as f32;
+        let new_atr = (prev_atr * (self.periods as f64 - 1.0) + tr) / self.periods as f64;
         self.value = Some(new_atr);
         self.prev_value = Some(new_atr);
     }
@@ -74,7 +74,7 @@ impl Indicator for Atr{
     fn update_before_close(&mut self, price: Price) {
     if let (Some(prev_close), Some(prev_atr)) = (self.prev_close, self.prev_value) {
         let tr = calc_tr(price.high, price.low, prev_close);
-        let provisional_atr = (prev_atr * (self.periods as f32 - 1.0) + tr) / self.periods as f32;
+        let provisional_atr = (prev_atr * (self.periods as f64 - 1.0) + tr) / self.periods as f64;
         self.value = Some(provisional_atr);
     }
 }
@@ -106,9 +106,9 @@ impl Indicator for Atr{
     }
 
 
-fn calc_tr(high: f32, low: f32, prev_close: f32) -> f32{
+fn calc_tr(high: f64, low: f64, prev_close: f64) -> f64{
 
-        f32::max(high - low, f32::max((high - prev_close).abs(), (low - prev_close).abs()))
+        f64::max(high - low, f64::max((high - prev_close).abs(), (low - prev_close).abs()))
     }
